Fix wrong path for authorized_keys

+++ entermediadb-sshd.sh
@@ -10,8 +10,11 @@ echo -e  'y\n' | ssh-keygen -f /etc/ssh/ssh_host_dsa_key -N '' -t dsa
 echo -e  'y\n' | ssh-keygen -A
 
 # Generate client SSH key with empty password
-ssh-keygen -t rsa -C "client@entermediadb.org" -f "/media/services/client.pk" -q -N ""
-mv /media/services/client.pk.pub /media/services/client.pub
+
+if [ ! -f /media/services/client.pk ]; then
+    ssh-keygen -t rsa -C "client@entermediadb.org" -f "/media/services/client.pk" -q -N ""
+	mv /media/services/client.pk.pub /media/services/client.pub
+fi
 
 # Create user client and set primary group to entermedia
 useradd client -g entermedia
@@ -19,6 +22,6 @@ useradd client -g entermedia
 #Create .ssh dir with required files
 mkdir $HOME_CLIENT/.ssh
 chmod -v 755 $HOME_CLIENT/.ssh
-cat /media/services/client.pub > $HOME_CLIENT/authorized_keys
-chmod -v 644 $HOME_CLIENT/authorized_keys
+cat /media/services/client.pub > $HOME_CLIENT/.ssh/authorized_keys
+chmod -v 644 $HOME_CLIENT/.ssh/authorized_keys
 chown -R client:entermedia $HOME_CLIENT/.ssh
